saving alterations and sorting works

+++ server/resolvers/Mutation/saveModification.js
@@ -1,6 +1,6 @@
 module.exports = async (
   parent,
-  { user, recipe, sortings, alterations },
+  { id, user, recipe, sortings, alterations, items },
   ctx
 ) => {
   const mod = await ctx.prisma
@@ -10,91 +10,42 @@ module.exports = async (
     .then(mods => mods.shift());
 
   if (mod) {
-    const sortedIds = await ctx.prisma
-      .modification({ id: mod.id })
-      .sortings()
-      .then(sortings => sortings.map(sorting => sorting.parentId));
-
-    const curAlterations = await ctx.prisma
-      .modification({ id: mod.id })
-      .alterations();
-
-    // Remove any modifications from the database
-    // which were removed on the front end
     await ctx.prisma.updateModification({
-      where: {
-        id: mod.id
-      },
+      where: { id: mod.id },
       data: {
         sortings: {
           deleteMany: {
-            parentId_not_in: sortings.map(sorting => sorting.parentId)
+            uid_not_in: sortings
+              .filter(sorting => sorting.uid)
+              .map(sorting => sorting.uid)
           }
         },
         alterations: {
           deleteMany: {
-            sourceId_not_in: alterations.map(alteration => alteration.sourceId),
-            field_not_in: alterations.map(alteration => alteration.field)
-            // NOT: alterations.map(alteration => ({
-            //   sourceId: alteration.sourceId,
-            //   field: alteration.field
-            // }))
+            uid_not_in: alterations
+              .filter(alteration => alteration.uid)
+              .map(alteration => alteration.uid)
           }
         }
       }
     });
 
-    // Update all existing modifications
-    await ctx.prisma.updateModification({
-      where: {
-        id: mod.id
-      },
+    return await ctx.prisma.updateModification({
+      where: { id: mod.id },
       data: {
         sortings: {
-          updateMany: sortings
-            .filter(sorting => sortedIds.includes(sorting.parentId))
+          update: sortings
+            .filter(sorting => sorting.uid)
             .map(sorting => ({
-              where: {
-                parentId: sorting.parentId
-              },
+              where: { uid: sorting.uid },
               data: {
                 order: {
                   set: sorting.order
                 }
               }
-            }))
-        },
-        alterations: {
-          updateMany: alterations
-            .filter(alteration =>
-              curAlterations.some(
-                curAlteration =>
-                  curAlteration.field === alteration.field &&
-                  curAlteration.sourceId === alteration.sourceId
-              )
-            )
-            .map(alteration => ({
-              where: {
-                sourceId: alteration.sourceId,
-                field: alteration.field
-              },
-              data: {
-                value: alteration.value
-              }
-            }))
-        }
-      }
-    });
-
-    // Create any new modifications
-    return await ctx.prisma.updateModification({
-      where: {
-        id: mod.id
-      },
-      data: {
-        sortings: {
+            })),
           create: sortings
-            .filter(sorting => !sortedIds.includes(sorting.parentId))
+            .filter(sorting => !sorting.uid)
             .map(sorting => ({
               parentId: sorting.parentId,
               order: {
@@ -103,15 +54,16 @@ module.exports = async (
             }))
         },
         alterations: {
+          update: alterations
+            .filter(alteration => alteration.uid)
+            .map(alteration => ({
+              where: { uid: alteration.uid },
+              data: {
+                value: alteration.value
+              }
+            })),
           create: alterations
-            .filter(
-              alteration =>
-                !curAlterations.some(
-                  curAlteration =>
-                    curAlteration.field === alteration.field &&
-                    curAlteration.sourceId === alteration.sourceId
-                )
-            )
+            .filter(alteration => !alteration.uid)
             .map(alteration => ({
               sourceId: alteration.sourceId,
               field: alteration.field,
@@ -121,7 +73,6 @@ module.exports = async (
       }
     });
   } else {
-    // Create a new modification and link it to the posted user
     return await ctx.prisma.createModification({
       recipe: {
         connect: { id: recipe }
